Added Sorting Support

+++ src/main/java/dslib/dataframe/kernels/IndexValueIntTuple.java
@@ -2,18 +2,18 @@ package dslib.dataframe.kernels;
 
 public class IndexValueIntTuple implements Comparable<IndexValueIntTuple> {
     private int index;
-    private int value;
+    private long value;
 
-    public IndexValueIntTuple(int index, int value) {
+    public IndexValueIntTuple(int index, long value) {
         this.index = index;
         this.value = value;
     }
 
     public int getIndex() { return index; }
-    public int getValue() { return value; }
+    public long getValue() { return value; }
 
     @Override
     public int compareTo(IndexValueIntTuple o) {
-        return this.value - o.value;
+        return (int)(this.value - o.value);
     }
 }
